backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
click==8.3.1
distro==1.9.0
fastapi==0.133.1
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
joblib==1.5.3
numpy==2.4.2
pandas==3.0.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.47
starlette==0.52.1
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
